requirements.txt
fastapi==0.112.2
uvicorn==0.30.1
httpx==0.27.0
pydantic==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
supabase-py==2.4.0

